Web_Clinic_4/flask/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
torch
transformers
mysqlclient
PyJWT
scipy
pymysql==1.0.2
cryptography